clean up new tag queries
